Retroactively allow sounddose V3 in 202404 FCM
There are changes in common audio libraries that force the sounddose
implementation to be bumped if any of the other audio implementations
are bumped.
Since we want to deprecate the sounddose interface and remove it from
FCMs > 202404, we need to retroactively allow the new sounddose version.
The other audio interfaces don't need this treatment because the newer
versions are added to the newer FCMs and we allow older devices to
implement them.

Test: m
Bug: 361383071
Change-Id: I195053bb6c5b6cb01584398eafc86f8fbcdd1148