libhealthloop: Reduce the size of the uevent receive buffer
The HealthLoop code only checks the name of the subsystem in the uevents
that it receives. Since the recently added BPF filter only passes power
supply uevents, all we need to know is whether or not any such uevents
have been received. We do not need to know how many of these events have
been received. Hence, reduce the size of the uevent receive buffer.
This CL reduces the number of ScheduleBatteryUpdate() calls if uevents
are received faster than these can be processed.

Bug: 362986353
Change-Id: If286ae5d05a95d59bc637a869c94d5c5472b5be2
Signed-off-by: Bart Van Assche <bvanassche@google.com>